pyproject.toml
[project]
name = "networkserver"
version = "0.1.0"
description = ""
authors = [
    {name = "avinash84319",email = "avihanvaisri@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pynacl (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
